requirements.txt
pillow
typing_extensions
facenet_pytorch
opencv-python
flask
IPython
timm
pygame
tqdm
ipywidgets